pyproject.toml
[project]
name = "csv2postgresql"
version = "0.1.0"
description = "Load CSV with header into PostgreSQL"
authors = [
    {name = "Francesco Frassinelli", email = "francesco.frassinelli@nina.no"},
]
dependencies = [
    "psycopg2-binary~=2.9",
]
requires-python = ">=3.7"
dynamic = ["classifiers"]
license = {text = "GPLv3.0+"}
readme = "README.md"

[project.urls]
homepage = "https://github.com/NINAnor/csv2postgresql"
repository = "https://github.com/NINAnor/csv2postgresql"

[project.optional-dependencies]

[project.scripts]
csv2postgresql = "csv2postgresql:cli"

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]

[tool.black]
line-length = 79

[tool.isort]
profile = "black"